requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.9.3
cffi==1.14.3
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
schedule==0.6.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
Werkzeug==1.0.1
